requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
Django==4.0.6
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.5.0
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.10
